requirements_updated.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncua==1.1.5
attrs==25.1.0
av==12.3.0
backoff==2.2.1
blinker==1.9.0
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
inference==0.38.0
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kornia==0.6.7
lazy_loader==0.4
loguru==0.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.27.1
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.570.86
onnxruntime==1.19.2
openai==1.63.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.7.0.72
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
Pillow==9.4.0
pluggy==1.5.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==5.29.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pybase64==1.0.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==2.0.0
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.11.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyreadline3==3.5.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
PyYAML==6.0
redis==5.0.8
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
rpds-py==0.22.3
s3transfer==0.10.4
scikit-image==0.24.0
scipy==1.15.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
slack_sdk==3.33.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
streamlit==1.42.1
structlog==24.4.0
supervision==0.25.1
sympy==1.13.3
tenacity==9.0.0
tifffile==2025.2.18
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==1.13.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.67.1
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
watchdog==6.0.0
win32_setctime==1.2.0
yarl==1.18.3
zxing-cpp==2.2.0
